agent/requirements.txt
agno>=1.7.8
openai>=1.88.0
yfinance>=0.2.63
fastapi>=0.115.13
uvicorn>=0.34.3
ag-ui-protocol>=0.1.8
packaging>=25.0.0
python-dotenv>=1.0.1
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.154.0
